:recycle: (TabLayout) Extract logic in hook
