+++ src/features/venue/components/TabLayout/useTabArrowNavigation.ts
@@ -1,0 +1,5 @@
+import { TabArrowNavigationProps } from 'features/venue/types'
+
+export const useTabArrowNavigation = (_: TabArrowNavigationProps): void => {
+  return
+}
